# swagger.yaml
swagger: '2.0'
info:
  description: Набор сервисов для учета на предприятии /ООО Ромашка/
  version: "1.0.0"
  title: RestApi

produces:
  - application/json
  - application/xml
  - application/text

basePath: "/api"

paths:
  /reports/formats:
    get:
      tags:
        - filter models
      summary: Получить список форматов для отчетов
      operationId: main.formats
      produces:
        - application/json
      responses:
        200:
          description: OK
          schema:
            type: object
        500:
          description: Ошибка на сервере
        default:
          description: Неизвестная ошибка     


  /reports/range/{format}:
    get:
      tags:
        - filter models
      summary: Получить список единиц измерения
      operationId: main.reports_range
      produces:
        - application/text
      responses:
        200:
          description: OK
          schema:
            type: object
        500:
          description: Ошибка на сервере
        default:
          description: Неизвестная ошибка
    parameters:
    - name: format
      in: path
      description: Уникальный код формата
      required: true
      type: string 

  /reports/group/{format}:
    get:
      tags:
        - filter models
      summary: Получить список групп номенклатур
      operationId: main.reports_group
      produces:
        - application/text
      responses:
        200:
          description: OK
          schema:
            type: object
        500:
          description: Ошибка на сервере
        default:
          description: Неизвестная ошибка
    parameters:
    - name: format
      in: path
      description: Уникальный код формата
      required: true
      type: string

  /reports/nomenclature/{format}:
    get:
      tags:
        - filter models
      summary: Получить список номенклатур
      operationId: main.reports_nomenclature
      produces:
        - application/text
      responses:
        200:
          description: OK
          schema:
            type: object
        500:
          description: Ошибка на сервере
        default:
          description: Неизвестная ошибка
    parameters:
    - name: format
      in: path
      description: Уникальный код формата
      required: true
      type: string

  /reports/recipe/{format}:
    get:
      tags:
        - filter models
      summary: Получить список рецептов
      operationId: main.reports_recipe
      produces:
        - application/text
      responses:
        200:
          description: OK
          schema:
            type: object
        500:
          description: Ошибка на сервере
        default:
          description: Неизвестная ошибка
    parameters:
    - name: format
      in: path
      description: Уникальный код формата
      required: true
      type: string

  /filter/{domain}:
      post:
        tags:
        - filter models
        summary: Фильтрация данных по модели
        operationId: main.filter_data
        consumes:
          - application/json
        produces:
          - application/json
        parameters:
          - name: domain
            in: path
            required: true
            type: string
            description: Доступные модели для фильтрации nomenclature, group, range, receipt
          - in: body
            name: filter
            description: Доступные фильтры EQUALS, LIKE
            required: true
            schema:
              type: object
              properties:
                name:
                  type: string
                  default: ""
                filter_name:
                  type: string
                  default: "EQUALS"
                unique_code:
                  type: string
                  default: ""
                filter_unique_code:
                  type: string
                  default: "EQUALS"
        responses:
          200:
            description: OK
            schema:
              type: object
          400:
            description: Ошибка ввода данных
            schema:
              type: object
          500:
            description: Ошибка на сервере
          default:
            description: Неизвестная ошибка


  /reports/transactions/{format}:
    get:
      tags:
        - filter warehouse
      summary: Получить отчет по транзакциям
      operationId: main.reports_transaction
      produces:
        - application/text
      responses:
        200:
          description: OK
          schema:
            type: object
        500:
          description: Ошибка на сервере
        default:
          description: Неизвестная ошибка
    parameters:
    - name: format
      in: path
      description: Уникальный код формата
      required: true
      type: string

  /reports/turnovers/{format}:
    get:
      tags:
        - filter warehouse
      summary: Получить отчет по оборотам
      operationId: main.reports_turnover
      produces:
        - application/text
      responses:
        200:
          description: OK
          schema:
            type: object
        500:
          description: Ошибка на сервере
        default:
          description: Неизвестная ошибка
    parameters:
    - name: format
      in: path
      description: Уникальный код формата
      required: true
      type: string
        

  /filter/transactions:
      post:
        tags:
          - filter warehouse
        summary: Фильтрация транзакций
        operationId: main.filter_transactions
        consumes:
          - application/json
        produces:
          - application/json
        parameters:
          - in: body
            name: filter
            description: Доступные модели для фильтрации - nomenclature, warehouse. Доступные фильтры EQUALS, LIKE
            required: true
            schema:
              type: object
              properties:
                warehouse: 
                  type: object
                  properties:
                    name:
                      type: string
                      default: ""
                    filter_name:
                      type: string
                      default: "EQUALS"
                    unique_code:
                      type: string
                      default: ""
                    filter_unique_code:
                      type: string
                      default: "EQUALS"
                nomenclature: 
                  type: object
                  properties:
                    name:
                      type: string
                      default: ""
                    filter_name:
                      type: string
                      default: "EQUALS"
                    unique_code:
                      type: string
                      default: ""
                    filter_unique_code:
                      type: string
                      default: "EQUALS"
        responses:
          200:
            description: OK
            schema:
              type: object
          400:
            description: Ошибка ввода данных
            schema:
              type: object
          500:
            description: Ошибка на сервере
          default:
            description: Неизвестная ошибка

  /block_period/get:
    get:
      tags:
        - block period
      summary: Просмотр текущей даты блокировки
      operationId: main.get_block_period
      produces:
        - application/text
      responses:
        200:
          description: OK
          schema:
            type: object
        500:
          description: Ошибка на сервере
        default:
          description: Неизвестная ошибка
  
  /block_period/set:
    post:
      tags:
        - block period
      summary: Изменение даты блокировки
      operationId: main.set_block_period
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - in: body
          name: set_block_period
          required: true
          schema:
            type: object
            properties:
              block_period:
                type: string
                format: date-time
                default: "2024-10-01"
      responses:
        200:
          description: OK
          schema:
            type: object
        400:
          description: Ошибка записи оборотов в файл
          schema:
            type: object
        404:
          description: Данных для фильтрации по модели нет.
          schema:
            type: object
        500:
          description: Неверный формат даты блокировки
        default:
          description: Неизвестная ошибка


  /api/nomenclature:
    get:
      tags:
        - nomenclature
      summary: Получить номенклатуру
      produces:
        - text/html
      operationId: main.get_nomenclature
      description: Возвращает информацию о номенклатуре по уникальному коду.
      parameters:
        - name: unique_code
          in: query
          description: Уникальный код номенклатуры
          required: true
          type: string
      responses:
        200:
          description: OK
        400:
          description: Ошибка ввода данных
        
        500:
          description: Неверный формат
        default:
          description: Неизвестная ошибка

    put:
      summary: Добавить номенклатуру
      tags:
        - nomenclature
      produces:
        - text/html
      operationId: main.put_nomenclature
      description: Создает новую номенклатуру.
      parameters:
        - in: body
          name: body
          description: Данные новой номенклатуры
          required: true
          schema:
            type: object
            properties:
              name:
                type: string
              group_id:
                type: string
              range_id:
                type: string
      responses:
        200:
          description: Номенклатура успешно создана!
        400:
          description: Ошибка ввода данных
        
        500:
          description: Неверный формат
        default:
          description: Неизвестная ошибка
    
    patch:
      summary: Обновить номенклатуру
      tags:
        - nomenclature
      produces:
        - text/html
      operationId: main.patch_nomenclature
      description: Обновляет существующую номенклатуру.
      parameters:
        - in: body
          name: body
          description: Данные для обновления номенклатуры
          required: true
          schema:
            type: object
            properties:
              unique_code:
                type: string
              name:
                type: string
              group_id:
                type: string
              range_id:
                type: string
      responses:
        200:
          description: OK
        400:
          description: Ошибка ввода данных
        
        500:
          description: Неверный формат
        default:
          description: Неизвестная ошибка

    delete:
      summary: Удалить номенклатуру
      tags:
        - nomenclature
      produces:
        - text/html
      operationId: main.delete_nomenclature
      description: Удаляет номенклатуру по уникальному коду.
      parameters:
        - name: unique_code
          in: query
          description: Уникальный код номенклатуры
          required: true
          type: string
      responses:
        200:
          description: OK
        400:
          description: Ошибка ввода данных
        
        500:
          description: Неверный формат
        default:
          description: Неизвестная ошибка

  
  /reports/balanse_sheet/{date_start}/{date_end}/{warehouse_id}:
    get:
      tags:
        - turnover balance sheet
      summary: Получить отчет ОСВ
      operationId: main.reports_balanse_sheet
      parameters:
        - name: date_start
          in: path
          description: Дата начала в формате 'YYYY-MM-DD'
          required: true
          type: string
        - name: date_end
          in: path
          description: Дата конца в формате 'YYYY-MM-DD'
          required: true
          type: string
        - name: warehouse_id
          in: path
          description: Код склада
          required: true
          type: string
      produces:
        - application/text
      responses:
        200:
          description: OK
          schema:
            type: object
        500:
          description: Ошибка на сервере
        default:
          description: Неизвестная ошибка


  /api/reposity/save:
    post:
      tags:
        - turnover balance sheet
      summary: Сохранить данные в файл
      operationId: main.save_reposity
      produces:
        - application/text
      responses:
        200:
          description: OK
          schema:
            type: object
        500:
          description: Ошибка на сервере
        default:
          description: Неизвестная ошибка

  /api/reposity/restore:
    post:
      tags:
        - turnover balance sheet
      summary: Загрузить данные из файла
      operationId: main.restore_reposity
      produces:
        - application/text
      responses:
        200:
          description: OK
          schema:
            type: object
        500:
          description: Ошибка на сервере
        default:
          description: Неизвестная ошибка
        